// frontend/src/lib/api/generated/models/gitChangesResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { PublishChange } from "./publishChange.ts";

export interface GitChangesResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  branch?: string;
  changes: PublishChange[];
  ignored_non_markdown_count: number;
  is_repo: boolean;
  suggested_message?: string;
  upstream?: string;
}