// frontend/src/lib/api/generated/models/publishResponse.ts
/**
 * Generated by Orval. Do not edit manually.
 */
import type { PublishChange } from "./publishChange.ts";

export interface PublishResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  branch: string;
  commit: string;
  files: PublishChange[];
  pushed: boolean;
  short_commit: string;
  upstream?: string;
}